Adding config validation and structed to packets.
