+++ 7th lesson/go.mod
@@ -1,5 +1,7 @@
-module github.com/pkMIXER/geek_go/2nd_lesson
+module github.com/pkMIXER/geek_go/7th_lesson
 
 go 1.15
 
-require gopkg.in/yaml.v2 v2.4.0 // indirect
+require (
+	gopkg.in/yaml.v2 v2.4.0
+)
